fix: isr 관련 오류 해결 시도
정적 생성 과정에서 쿠키를 사용하지 않는 방법으로 스코프 오류 해결 시도
Related to: #44
